+++ site/pages/index.jsx
@@ -7,8 +7,7 @@ export default {
             'layout': require('../../src/components/layout/docs'),
             'container': require('../../src/components/container/docs'),
             'button': require('../../src/components/button/docs'),
-            'breadcrumb': require('../../src/components/breadcrumb/docs'),
-            'select': require('../../src/components/select/docs')
+            'breadcrumb': require('../../src/components/breadcrumb/docs')
         },
         'Dados': {
             'tag': require('../../src/components/tag/docs')
@@ -20,8 +19,9 @@ export default {
         'Formulários': {
             'input': require('../../src/components/input/docs'),
             'form': require('../../src/components/form/docs'),
-            'radio': require('../../src/components/radio/docs'),
-            'table': require('../../src/components/table/docs')
+            'table': require('../../src/components/table/docs'),
+            'select': require('../../src/components/select/docs'),
+            'radio': require('../../src/components/radio/docs')
         },
         'Outros': {
             'card': require('../../src/components/card/docs'),