+++ src/components/select/Select.test.jsx
@@ -60,7 +60,7 @@ const options = [
     }
 ];
 
-let value = ''
+let value = '';
 
 
 describe('Select test', () => {
@@ -74,13 +74,14 @@ describe('Select test', () => {
                     })
                 }
             </Select>
-        )
+        );
 
         expect(select.find('.cd-select').exists()).toBe(true);
         expect(select.find('ul').children().length).toBe(5);
         select.find('Input').simulate('click');
         select.find('ul').childAt(1).find('li').simulate('click');
         expect(select.find('Select').state().selectedLabel).toBe('Option2')
+        select.find('.cd-input__inner').simulate('change', { target: { value: 'My new value' } });
 
     });
 
@@ -93,7 +94,7 @@ describe('Select test', () => {
                     })
                 }
             </Select>
-        )
+        );
 
         expect(select.find('.cd-select').exists()).toBe(true);
         expect(select.find('.cd-icon-time').exists()).toBe(true);
@@ -109,7 +110,7 @@ describe('Select test', () => {
                     })
                 }
             </Select>
-        )
+        );
 
         expect(select.find('.cd-select').exists()).toBe(true);
         expect(select.find('Input').childAt(0).hasClass('is-disabled')).toBe(true);
@@ -129,14 +130,26 @@ describe('Select test', () => {
                     })
                 }
             </Select>
-        )
+        );
+
+        let clickEvent = new MouseEvent('click');
 
         expect(select.find('.cd-select').exists()).toBe(true);
         expect(select.find('ul').childAt(1).find('li').hasClass('is-disabled')).toBe(true);
         select.find('ul').childAt(1).find('li').simulate('click');
-        expect(select.find('Select').state().selectedLabel).toBe('')
+        select.find('.cd-scrollbar__bar.is-vertical').simulate('mousedown');
+        select.find('.cd-scrollbar__wrap').simulate('scroll');
+        select.find('Scrollbar').instance()._update();
+        select.find('Bar').first().instance().mouseMoveDocumentHandler(clickEvent)
+        select.find('Bar').first().instance().mouseUpDocumentHandler(clickEvent)
+        select.find('.cd-scrollbar__bar.is-vertical')
+              .find('.cd-scrollbar__thumb')
+              .simulate('mousedown');
+        expect(select.find('Select').state().selectedLabel).toBe('');
         select.find('ul').childAt(2).find('li').simulate('click');
         expect(select.find('Select').state().selectedLabel).toBe('Option3')
+
+        select.unmount()
 
     });
 
@@ -150,12 +163,12 @@ describe('Select test', () => {
                     })
                 }
             </Select>
-        )
+        );
 
         expect(select.find('.cd-select').exists()).toBe(true);
         select.find('Input').simulate('click');
         select.find('ul').childAt(1).find('li').simulate('click');
-        expect(select.find('Select').state().selectedLabel).toBe('Option2')
+        expect(select.find('Select').state().selectedLabel).toBe('Option2');
         select.find('Input').simulate('mouseenter');
         select.find('.cd-icon-circle-close').simulate('click');
         expect(fn.callCount).toBe(1);
@@ -172,15 +185,16 @@ describe('Select test', () => {
                     })
                 }
             </Select>
-        )
+        );
 
         expect(select.find('.cd-select').exists()).toBe(true);
         select.find('Input').simulate('click');
         select.find('ul').childAt(1).find('li').simulate('click');
         select.find('ul').childAt(2).find('li').simulate('click');
-        expect(Array.isArray(select.find('Select').state().selectedLabel)).toBe(true)
-        expect(select.find('Select').state().selectedLabel[0]).toBe('Option2')
+        expect(Array.isArray(select.find('Select').state().selectedLabel)).toBe(true);
+        expect(select.find('Select').state().selectedLabel[0]).toBe('Option2');
         expect(select.find('Select').state().selectedLabel[1]).toBe('Option3')
+        select.find('Select').find('.cd-icon-close').first().simulate('click')
 
     });
 
@@ -202,9 +216,12 @@ describe('Select test', () => {
                     })
                 }
             </Select>
-        )
+        );
 
         expect(select.find('.cd-select').exists()).toBe(true);
+        select.find('Input').simulate('mouseenter')
+        select.find('Input').simulate('mouseleave')
+        select.find('Input').simulate('mousedown')
         expect(select.find('ul').children().length).toBe(5);
 
     });
@@ -238,7 +255,7 @@ describe('Select test', () => {
                     })
                 }
             </Select>
-        )
+        );
 
         expect(select.find('.cd-select').exists()).toBe(true);
         expect(select.find('ul').children().length).toBe(12);
@@ -254,19 +271,44 @@ describe('Select test', () => {
                     })
                 }
             </Select>
-        )
+        );
 
         expect(select.find('.cd-select').exists()).toBe(true);
+        select.find('Select').simulate('click');
+
         expect(select.find('ul').children().length).toBe(5);
 
+        jest.useFakeTimers()
+        select.find('Input').find('input').simulate('click');
+        jest.runAllTimers()
+        select.find('.cd-input__inner').simulate('change', { target: { value: 'My new value' } });
+
+        select.find('Select').setState({
+                                           query: 'Option1'
+                                       });
+
+        expect(select.find('Select').state().filteredOptionsCount).toBe(1)
+
+        select.find('.cd-input__inner').simulate('keypress', { keyCode: 27 })
+        select.find('.cd-input__inner').simulate('keydown', { keyCode: 27 })
+        select.find('.cd-input__inner').simulate('keypress', { keyCode: 13 })
+        select.find('.cd-input__inner').simulate('keydown', { keyCode: 13 })
+        select.find('.cd-input__inner').simulate('keypress', { keyCode: 8 })
+        select.find('.cd-input__inner').simulate('keydown', { keyCode: 8 })
+        select.find('.cd-input__inner').simulate('keypress', { keyCode: 38 })
+        select.find('.cd-input__inner').simulate('keydown', { keyCode: 38 })
+        select.find('.cd-input__inner').simulate('keypress', { keyCode: 40 })
+        select.find('.cd-input__inner').simulate('keydown', { keyCode: 40 })
+        select.find('.cd-input__inner').simulate('keyup')
+        select.find('.cd-input__inner').simulate('mousedown')
     });
 
-    /*it('remote', () => {
+    it('remote', () => {
 
         let loading = false;
 
         const select = mount(
-          <Select multiple filterable remote loading={loading}>
+            <Select value={value} multiple filterable remote loading={loading}>
             {
                     options.map(el => {
                         return <Option key={el.value} label={el.label} value={el.value} />
@@ -276,7 +318,81 @@ describe('Select test', () => {
         )
 
         expect(select.find('.cd-select').exists()).toBe(true);
-        expect(select.find('ul').children().length).toBe(5);
 
-    });*/
+        expect(select.find('Select').state().visible).toBe(false)
+
+        select.find('Input').find('input').simulate('click');
+
+        expect(select.find('Select').state().visible).toBe(true)
+
+        select.find('.cd-input__inner').simulate('change', { target: { value: 'My new value' } });
+
+        select.find('ul').childAt(1).find('li').simulate('mouseenter');
+
+        select.find('Input').find('input').simulate('click');
+
+        select.find('ul').childAt(1).find('li').simulate('mouseleave');
+
+        select.find('Select').setState({
+                                           query: 'Option1'
+                                       });
+
+        select.find('Input').simulate('mousedown')
+
+        expect(select.find('Select').state().filteredOptionsCount).toBe(1)
+
+        select.find('Select').instance().handleClickOutside()
+
+        expect(select.find('Select').state().visible).toBe(false)
+
+        select.find('.cd-select__input').simulate('keypress', { keyCode: 27 })
+        select.find('.cd-select__input').simulate('keydown', { keyCode: 27 })
+        select.find('.cd-select__input').simulate('keypress', { keyCode: 13 })
+        select.find('.cd-select__input').simulate('keydown', { keyCode: 13 })
+        select.find('.cd-select__input').simulate('keypress', { keyCode: 8 })
+        select.find('.cd-select__input').simulate('keydown', { keyCode: 8 })
+        select.find('.cd-select__input').simulate('keypress', { keyCode: 38 })
+        select.find('.cd-select__input').simulate('keydown', { keyCode: 38 })
+        select.find('.cd-select__input').simulate('keypress', { keyCode: 40 })
+        select.find('.cd-select__input').simulate('keydown', { keyCode: 40 })
+        select.find('.cd-select__input').simulate('keyup')
+        jest.useFakeTimers();
+        select.find('.cd-select__input').simulate('change', { target: { value: 'My new value' } })
+        jest.runAllTimers();
+
+    });
+
+
+    it('remote with method', () => {
+
+        let loading = false;
+
+        const select = mount(
+            <Select value="Option1" remoteMethod={() => {
+            }} multiple filterable remote loading={loading}>
+                {
+                    options.map(el => {
+                        return <Option key={el.value} label={el.label} value={el.value}/>
+                    })
+                }
+            </Select>
+        )
+        select.setProps({
+                            placeholder: 'placeholder',
+                            value: ['Option1', 'Option2'],
+                            onVisibleChange: () => {
+                            }
+                        });
+
+        select.find('Select').setState({
+                                           query: 'Option1',
+                                           visible: true
+                                       });
+
+        select.find('Input').find('input').simulate('click');
+
+        select.find('Select').instance().handleClickOutside()
+
+    });
+
 });